refactor(io): split training_history into recorder + training_history, real epoch in record()
- io/recorder.rs: FileSnapshotRecorder + TrainingMeta (write side)
- io/training_history.rs: TrainingHistory::load + model_at (read side)
- SnapshotRecorder::record gains epoch: usize — caller passes the real
  training epoch; FileSnapshotRecorder no longer computes count*interval
- FileSnapshotRecorder drops interval from its struct and from resume()
- TrainingMeta renamed from SnapshotMeta (describes the training run, not a snapshot)
- nn/recorders/ split into mod.rs (trait) + checkpoints.rs + noop.rs
- io/recorder.rs renamed from io/snapshot_recorder.rs
